+++ spec/javascripts/collectionView.reset.spec.js
@@ -56,7 +56,12 @@ describe("collection view - reset", function(){
     });
 
     it("should remove the event handlers for the original children", function(){
-      expect(_.size(collectionView._listeners)).toBe(4);
+      // maintain backwards compatibility with backbone 1.0.0 in tests
+      if (typeof collectionView._listeningTo != "undefined") {
+        expect(_.size(collectionView._listeningTo)).toBe(4);
+      } else {
+        expect(_.size(collectionView._listeners)).toBe(4);
+      }
     });
   });