+++ spec/javascripts/templateCache.spec.js
@@ -9,7 +9,7 @@ describe("template cache", function(){
 
       Backbone.Marionette.TemplateCache.get("#t1");
     });
-    
+
     it("should load from the DOM", function(){
       expect(Backbone.Marionette.TemplateCache.prototype.loadTemplate).toHaveBeenCalled();
     });
@@ -30,7 +30,7 @@ describe("template cache", function(){
       Backbone.Marionette.TemplateCache.get("#t2");
       Backbone.Marionette.TemplateCache.get("#t2");
     });
-    
+
     it("should load from the DOM once", function(){
       expect(templateCache.loadTemplate).not.toHaveBeenCalled();
       expect(templateCache.loadTemplate.callCount).toBe(0);
@@ -44,7 +44,7 @@ describe("template cache", function(){
 
       Backbone.Marionette.TemplateCache.clear();
     });
-    
+
     it("should clear the cache", function(){
       expect(_.size(Backbone.Marionette.TemplateCache.templateCaches)).toBe(0);
     });
@@ -59,7 +59,7 @@ describe("template cache", function(){
 
       Backbone.Marionette.TemplateCache.clear("#t4");
     });
-    
+
     it("should clear the specified templates cache", function(){
       expect(Backbone.Marionette.TemplateCache.templateCaches["#t4"]).toBeUndefined();
     });
@@ -79,7 +79,7 @@ describe("template cache", function(){
 
       Backbone.Marionette.TemplateCache.clear("#t4", "#t5");
     });
-    
+
     it("should clear the specified templates cache", function(){
       expect(Backbone.Marionette.TemplateCache.templateCaches["#t4"]).toBeUndefined();
       expect(Backbone.Marionette.TemplateCache.templateCaches["#t5"]).toBeUndefined();