+++ spec/javascripts/view.spec.js
@@ -130,6 +130,43 @@ describe("base view", function(){
     });
   });
 
+  describe("constructing a view with default options", function(){
+    var view = Marionette.ItemView.extend();
+    var presetOptions = Marionette.View.extend({
+      options: {
+        'lila': 'zoidberg'
+      }
+    });
+
+    it("should take and store view options", function() {
+      var viewInstance = new view({"Guybrush": "Island"});
+      expect(viewInstance.options.Guybrush).toBe("Island");
+    });
+
+    it("should have an empty hash of options by default", function() {
+      var viewInstance = new view;
+      expect(typeof(viewInstance.options.Guybrush)).toBe("undefined");
+    });
+
+    it("should retain options set on view class", function() {
+      var viewInstance = new presetOptions;
+      expect(viewInstance.options.lila).toBe("zoidberg");
+    });
+  });
+
+  describe("should expose its options in the constructor", function() {
+    var View = Marionette.View.extend({
+       initialize: function() {
+        this.info = this.options;
+      }
+    });
+
+    it("should be able to access instance options", function() {
+      var myView = new View({name: "LeChuck"});
+      expect(myView.info.name).toBe("LeChuck");
+    });
+  });
+
   describe("when closing a view that is already closed", function(){
     var close, view;