+++ spec/javascripts/view.triggers.spec.js
@@ -137,4 +137,93 @@ describe("view triggers", function(){
     });
   });
 
+  describe("triggers should stop propigation and events by default", function() {
+    var myView = Backbone.Marionette.ItemView.extend({
+      triggers: {
+        'click h2': "headline:clicked"
+      },
+
+      initialize: function() {
+        this.spanClicked = false;
+      },
+
+      onRender: function() {
+        var self = this;
+        this.$('span').on("click", function() {
+          self.spanClicked = true;
+        });
+      },
+
+      template: _.template("<h2><span>hi</span></h2><a href='#hash-url'>hash link</a>")
+    });
+
+    var viewInstance;
+
+    beforeEach(function(){
+      viewInstance = new myView;
+      spyOn(window, 'onhashchange');
+
+      viewInstance.render();
+      viewInstance.$('h2').click();
+      viewInstance.$('a').click();
+    });
+
+    it("should stop propigation by default", function(){
+      expect(viewInstance.spanClicked).toBe(false);
+    });
+
+    it("should prevent default by default", function() {
+      expect(window.onhashchange).not.toHaveBeenCalled();
+    });
+  });
+
+  describe("when triggers items are manually configured", function(){
+    var View = Backbone.Marionette.ItemView.extend({
+      triggers: {
+        "click .foo": {
+          event: "do:foo",
+          preventDefault: true,
+          stopPropagation: false
+        },
+        "click .bar": {
+          event: "do:bar",
+          stopPropagation: true
+        }
+      },
+
+      render: function(){
+        this.$el.html("<button class='foo'></button><a href='#' class='bar'>asdf</a>");
+      }
+    });
+
+    var view, fooEvent, barEvent;
+
+    beforeEach(function(){
+      view = new View();
+      view.render();
+
+      fooEvent = $.Event("click");
+      barEvent = $.Event("click");
+
+      spyOn(fooEvent, 'preventDefault');
+      spyOn(fooEvent, 'stopPropagation');
+
+      spyOn(barEvent, 'preventDefault');
+      spyOn(barEvent, 'stopPropagation');
+
+      view.$(".foo").trigger(fooEvent);
+      view.$(".bar").trigger(barEvent);
+    });
+
+    it("should prevent and don't stop the first view event", function(){
+      expect(fooEvent.preventDefault).toHaveBeenCalled();
+      expect(fooEvent.stopPropagation).not.toHaveBeenCalled();
+    });
+
+    it("should not prevent and stop the second view event", function(){
+      expect(barEvent.preventDefault).not.toHaveBeenCalled();
+      expect(barEvent.stopPropagation).toHaveBeenCalled();
+    });
+  });
+
 });